+++ packages/react/src/components/form/Combobox/internal/ComboboxChips.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 
-import { ChipRemovable } from '../../Chip';
+import { ChipRemovable } from '../../../Chip';
 import { ComboboxContext } from '../Combobox';
 
 export const ComboboxChips = () => {
